lavc/vvc: Change intra pixel type to `int16_t`
Previously the pixels were stored as `int`s. The maximum bit depth
supported by VVC is 16, hence an `int16_t` is guaranteed to be
sufficient. Smaller sizes allow for more pixels to be packed into an
SIMD register and therefore faster assembly optimisations. Further
optimisation could be achieved by using `int8_t`s if the bit depth of
the current video is only 8.