lavc/vvc: Output ITX result in column-major order
This is laying groundwork for assembly optimisations. It is easier to
perform a column transform using SIMD instructions than it is to perform
a column transform. Typically, a transposition is required. Therefore,
by expecting the output of the inverse transforms in column-major order,
a transposition can be eliminated. This is also effectively free, with
only a minor performance hit due to cache-friendliness.