vvc_deblock: re-arrange spatial activity
spatial activity will use P1 instead of P2/P3 if max_len_p = 1 but only for determining the spatial activity; the actual calculations still use P2/P3 and need to be reloaded.

We have to move this after jumping to weak (assuming no strong), this previously wasn't a problem because we would also skip the store back to the memory addresses. However to prep for vertical we need to transpose the entire set of lines

squash